refactor to use page layout
