+++ apps/next/pages/_app.tsx
@@ -10,6 +10,9 @@ import { NextThemeProvider, useRootTheme } from '@tamagui/next-theme'
 import type { SolitoAppProps } from 'solito'
 
 function MyApp({ Component, pageProps }: SolitoAppProps) {
+  // Use the layout defined at the page level, if available
+  const getLayout = Component.getLayout || ((page) => page)
+
   return (
     <>
       <Head>
@@ -17,9 +20,7 @@ function MyApp({ Component, pageProps }: SolitoAppProps) {
         <meta name="description" content="One chat app to rule them all" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <ThemeProvider>
-        <Component {...pageProps} />
-      </ThemeProvider>
+      <ThemeProvider>{getLayout(<Component {...pageProps} />)}</ThemeProvider>
     </>
   )
 }
@@ -39,3 +40,14 @@ function ThemeProvider({ children }: { children: React.ReactNode }) {
 }
 
 export default MyApp
+
+// import Layout from '../components/layout'
+
+// const MyApp = ({ Component, pageProps }) => {
+//   // Use the layout defined at the page level, if available
+//   const getLayout = Component.getLayout || ((page) => page)
+
+//   return getLayout(<Component {...pageProps} />)
+// }
+
+// export default MyApp
